requirements.txt
flask
requests
openai
deepgram-sdk
python-dotenv